+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb.js
@@ -214,6 +214,50 @@ function deserialize_api_container_api_StartServiceResponse(buffer_arg) {
   return api_container_service_pb.StartServiceResponse.deserializeBinary(new Uint8Array(buffer_arg));
 }
 
+function serialize_api_container_api_StoreFilesArtifactFromServiceArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.StoreFilesArtifactFromServiceArgs)) {
+    throw new Error('Expected argument of type api_container_api.StoreFilesArtifactFromServiceArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_StoreFilesArtifactFromServiceArgs(buffer_arg) {
+  return api_container_service_pb.StoreFilesArtifactFromServiceArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_StoreFilesArtifactFromServiceResponse(arg) {
+  if (!(arg instanceof api_container_service_pb.StoreFilesArtifactFromServiceResponse)) {
+    throw new Error('Expected argument of type api_container_api.StoreFilesArtifactFromServiceResponse');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_StoreFilesArtifactFromServiceResponse(buffer_arg) {
+  return api_container_service_pb.StoreFilesArtifactFromServiceResponse.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_StoreWebFilesArtifactArgs(arg) {
+  if (!(arg instanceof api_container_service_pb.StoreWebFilesArtifactArgs)) {
+    throw new Error('Expected argument of type api_container_api.StoreWebFilesArtifactArgs');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_StoreWebFilesArtifactArgs(buffer_arg) {
+  return api_container_service_pb.StoreWebFilesArtifactArgs.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
+function serialize_api_container_api_StoreWebFilesArtifactResponse(arg) {
+  if (!(arg instanceof api_container_service_pb.StoreWebFilesArtifactResponse)) {
+    throw new Error('Expected argument of type api_container_api.StoreWebFilesArtifactResponse');
+  }
+  return Buffer.from(arg.serializeBinary());
+}
+
+function deserialize_api_container_api_StoreWebFilesArtifactResponse(buffer_arg) {
+  return api_container_service_pb.StoreWebFilesArtifactResponse.deserializeBinary(new Uint8Array(buffer_arg));
+}
+
 function serialize_api_container_api_UnloadModuleArgs(arg) {
   if (!(arg instanceof api_container_service_pb.UnloadModuleArgs)) {
     throw new Error('Expected argument of type api_container_api.UnloadModuleArgs');
@@ -463,7 +507,7 @@ getModules: {
     responseSerialize: serialize_api_container_api_GetModulesResponse,
     responseDeserialize: deserialize_api_container_api_GetModulesResponse,
   },
-  // Uploads a file to the Kurtosis File System.
+  // Uploads a files artifact to the Kurtosis File System.
 uploadFilesArtifact: {
     path: '/api_container_api.ApiContainerService/UploadFilesArtifact',
     requestStream: false,
@@ -474,6 +518,30 @@ uploadFilesArtifact: {
     requestDeserialize: deserialize_api_container_api_UploadFilesArtifactArgs,
     responseSerialize: serialize_api_container_api_UploadFilesArtifactResponse,
     responseDeserialize: deserialize_api_container_api_UploadFilesArtifactResponse,
+  },
+  // Tells the API container to download a files artifact from the web to the Kurtosis File System
+storeWebFilesArtifact: {
+    path: '/api_container_api.ApiContainerService/StoreWebFilesArtifact',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.StoreWebFilesArtifactArgs,
+    responseType: api_container_service_pb.StoreWebFilesArtifactResponse,
+    requestSerialize: serialize_api_container_api_StoreWebFilesArtifactArgs,
+    requestDeserialize: deserialize_api_container_api_StoreWebFilesArtifactArgs,
+    responseSerialize: serialize_api_container_api_StoreWebFilesArtifactResponse,
+    responseDeserialize: deserialize_api_container_api_StoreWebFilesArtifactResponse,
+  },
+  // Tells the API container to copy a files artifact from a service to the Kurtosis File System
+storeFilesArtifactFromService: {
+    path: '/api_container_api.ApiContainerService/StoreFilesArtifactFromService',
+    requestStream: false,
+    responseStream: false,
+    requestType: api_container_service_pb.StoreFilesArtifactFromServiceArgs,
+    responseType: api_container_service_pb.StoreFilesArtifactFromServiceResponse,
+    requestSerialize: serialize_api_container_api_StoreFilesArtifactFromServiceArgs,
+    requestDeserialize: deserialize_api_container_api_StoreFilesArtifactFromServiceArgs,
+    responseSerialize: serialize_api_container_api_StoreFilesArtifactFromServiceResponse,
+    responseDeserialize: deserialize_api_container_api_StoreFilesArtifactFromServiceResponse,
   },
 };