+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_grpc_web_pb.js
@@ -1048,5 +1048,127 @@ proto.api_container_api.ApiContainerServicePromiseClient.prototype.uploadFilesAr
 };
 
 
+/**
+ * @const
+ * @type {!grpc.web.MethodDescriptor<
+ *   !proto.api_container_api.StoreWebFilesArtifactArgs,
+ *   !proto.api_container_api.StoreWebFilesArtifactResponse>}
+ */
+const methodDescriptor_ApiContainerService_StoreWebFilesArtifact = new grpc.web.MethodDescriptor(
+  '/api_container_api.ApiContainerService/StoreWebFilesArtifact',
+  grpc.web.MethodType.UNARY,
+  proto.api_container_api.StoreWebFilesArtifactArgs,
+  proto.api_container_api.StoreWebFilesArtifactResponse,
+  /**
+   * @param {!proto.api_container_api.StoreWebFilesArtifactArgs} request
+   * @return {!Uint8Array}
+   */
+  function(request) {
+    return request.serializeBinary();
+  },
+  proto.api_container_api.StoreWebFilesArtifactResponse.deserializeBinary
+);
+
+
+/**
+ * @param {!proto.api_container_api.StoreWebFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>} metadata User defined
+ *     call metadata
+ * @param {function(?grpc.web.RpcError, ?proto.api_container_api.StoreWebFilesArtifactResponse)}
+ *     callback The callback function(error, response)
+ * @return {!grpc.web.ClientReadableStream<!proto.api_container_api.StoreWebFilesArtifactResponse>|undefined}
+ *     The XHR Node Readable Stream
+ */
+proto.api_container_api.ApiContainerServiceClient.prototype.storeWebFilesArtifact =
+    function(request, metadata, callback) {
+  return this.client_.rpcCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/StoreWebFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_StoreWebFilesArtifact,
+      callback);
+};
+
+
+/**
+ * @param {!proto.api_container_api.StoreWebFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>=} metadata User defined
+ *     call metadata
+ * @return {!Promise<!proto.api_container_api.StoreWebFilesArtifactResponse>}
+ *     Promise that resolves to the response
+ */
+proto.api_container_api.ApiContainerServicePromiseClient.prototype.storeWebFilesArtifact =
+    function(request, metadata) {
+  return this.client_.unaryCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/StoreWebFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_StoreWebFilesArtifact);
+};
+
+
+/**
+ * @const
+ * @type {!grpc.web.MethodDescriptor<
+ *   !proto.api_container_api.StoreFilesArtifactFromServiceArgs,
+ *   !proto.api_container_api.StoreFilesArtifactFromServiceResponse>}
+ */
+const methodDescriptor_ApiContainerService_StoreFilesArtifactFromService = new grpc.web.MethodDescriptor(
+  '/api_container_api.ApiContainerService/StoreFilesArtifactFromService',
+  grpc.web.MethodType.UNARY,
+  proto.api_container_api.StoreFilesArtifactFromServiceArgs,
+  proto.api_container_api.StoreFilesArtifactFromServiceResponse,
+  /**
+   * @param {!proto.api_container_api.StoreFilesArtifactFromServiceArgs} request
+   * @return {!Uint8Array}
+   */
+  function(request) {
+    return request.serializeBinary();
+  },
+  proto.api_container_api.StoreFilesArtifactFromServiceResponse.deserializeBinary
+);
+
+
+/**
+ * @param {!proto.api_container_api.StoreFilesArtifactFromServiceArgs} request The
+ *     request proto
+ * @param {?Object<string, string>} metadata User defined
+ *     call metadata
+ * @param {function(?grpc.web.RpcError, ?proto.api_container_api.StoreFilesArtifactFromServiceResponse)}
+ *     callback The callback function(error, response)
+ * @return {!grpc.web.ClientReadableStream<!proto.api_container_api.StoreFilesArtifactFromServiceResponse>|undefined}
+ *     The XHR Node Readable Stream
+ */
+proto.api_container_api.ApiContainerServiceClient.prototype.storeFilesArtifactFromService =
+    function(request, metadata, callback) {
+  return this.client_.rpcCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/StoreFilesArtifactFromService',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_StoreFilesArtifactFromService,
+      callback);
+};
+
+
+/**
+ * @param {!proto.api_container_api.StoreFilesArtifactFromServiceArgs} request The
+ *     request proto
+ * @param {?Object<string, string>=} metadata User defined
+ *     call metadata
+ * @return {!Promise<!proto.api_container_api.StoreFilesArtifactFromServiceResponse>}
+ *     Promise that resolves to the response
+ */
+proto.api_container_api.ApiContainerServicePromiseClient.prototype.storeFilesArtifactFromService =
+    function(request, metadata) {
+  return this.client_.unaryCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/StoreFilesArtifactFromService',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_StoreFilesArtifactFromService);
+};
+
+
 module.exports = proto.api_container_api;