+++ core/api/typescript/src/lib/enclaves/generic_api_container_client.ts
@@ -22,7 +22,7 @@ import {
     RemoveServiceArgs,
     RepartitionArgs,
     StartServiceArgs,
-    StartServiceResponse,
+    StartServiceResponse, StoreFilesArtifactFromServiceArgs, StoreWebFilesArtifactArgs, StoreWebFilesArtifactResponse,
     UnloadModuleArgs,
     UploadFilesArtifactArgs,
     UploadFilesArtifactResponse,
@@ -49,4 +49,6 @@ export interface GenericApiContainerClient {
     executeModule(executeModuleArgs: ExecuteModuleArgs): Promise<Result<ExecuteModuleResponse, Error>>
     execCommand(execCommandArgs: ExecCommandArgs): Promise<Result<ExecCommandResponse, Error>>
     uploadFiles(uploadFilesArtifactArgs: UploadFilesArtifactArgs): Promise<Result<UploadFilesArtifactResponse, Error>>
+    storeWebFilesArtifact(storeWebFilesArtifactArgs: StoreWebFilesArtifactArgs): Promise<Result<StoreWebFilesArtifactResponse, Error>>
+    storeFilesArtifactFromService(storeFilesArtifactFromServiceArgs: StoreFilesArtifactFromServiceArgs): Promise<Result<StoreWebFilesArtifactResponse, Error>>
 }